Restructuring the files and directories
