+++ src/views/Home.vue
@@ -1,23 +1,23 @@
-<template lang="pug">
+<template lang='pug'>
   div
     nav-bar
-    b-container-fluid(align-v="center")
-      b-jumbotron(header="Looking for Notes, Past Papers?", lead="Find everything here.")
+    b-container-fluid(align-v='center')
+      b-jumbotron(header='Looking for Notes, Past Papers?', lead='Find everything here.')
         b-input-group
           b-form-input#search-box.align-self-center
         p#search-btn
-          b-button(variant="outline-success") Search
+          b-button(variant='outline-success') Search
     page-footer
 </template>
 
 <script>
-import NavBar from "@/subcomponents/NavBar";
-import Footer from "@/subcomponents/Footer";
+// import Header from '@/components/TheHeader';
+import Footer from '@/components/TheFooter';
 
 export default {
-  name: "Home",
+  name: 'Home',
   components: {
-    'nav-bar': NavBar,
+    // 'page-header': Header,
     'page-footer': Footer
   }
 }
